An error was occurring when deserializing large numbers out of range …

+++ src/main/java/com/trustedchoice/askkodiak/v2/model/Range.java
@@ -28,7 +28,7 @@ package com.trustedchoice.askkodiak.v2.model;
 @lombok.Data
 public class Range {
 
-    private Integer max;
+    private Long max;
 
-    private Integer min;
+    private Long min;
 }
